refactor pyspv -> HornNode
