+++ pytxhorn/contrib/core/monitors/stealth.py
